FIX: cleanner version

+++ inc/parser/parser.h
@@ -6,7 +6,7 @@
 /*   By: yboudoui <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/16 13:47:36 by yboudoui          #+#    #+#             */
-/*   Updated: 2023/01/21 12:22:08 by yboudoui         ###   ########.fr       */
+/*   Updated: 2023/01/22 16:53:18 by yboudoui         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,13 +14,14 @@
 # define PARSER_H
 
 # include "utils.h"
+# include <stdint.h>
 
 typedef struct s_input {
-	int	number_of_philosophers;
-	int	time_to_die;
-	int	time_to_eat;
-	int	time_to_sleep;
-	int	number_of_times_each_philosopher_must_eat;
+	uint64_t	number_of_philosophers;
+	uint64_t	time_to_die;
+	uint64_t	time_to_eat;
+	uint64_t	time_to_sleep;
+	uint64_t	number_of_times_each_philosopher_must_eat;
 }	t_input;
 
 int	parse_arg(int ac, char *av[], t_input *out);
